Split MessageFactory to ClientMessageFactory and ServerMessageFactory
